SO 4178-6918: Add qs-main.c and consequential changes
