+++ src/so-4178-6918/qs-main.c
@@ -1,0 +1,40 @@
+/* SO 4178-6918 */
+/* Common main program */
+
+#include "qs-mon.h"
+#include <stdbool.h>
+#include <stdio.h>
+#include <string.h>
+
+int main(void)
+{
+    int data[] =
+    {
+        /* random -n 20 0 9 | commalist -b '        ' -n 10 */
+        //4, 8, 0, 0, 3, 8, 3, 6, 5, 9,
+        //6, 8, 3, 6, 5, 5, 0, 8, 1, 1,
+        //3, 9, 4, 7, 2, 6, 9, 0, 6, 1,
+        //8, 0, 2, 1, 4, 0, 6, 5, 4, 2,
+        //7, 6, 2, 5, 4, 4, 6, 0, 8, 3,
+        //6, 1, 2, 7, 4, 3, 0, 0, 0, 4,
+        //4, 7, 8, 8, 4, 4, 4, 4, 9, 6,
+        //9, 0, 2, 7, 6, 5, 9, 2, 7, 7,
+        9, 7, 0, 9, 5, 4, 8, 7, 9, 9,
+        2, 9, 9, 7, 0, 3, 9, 6, 8, 5,
+        //5, 1, 4, 5, 5, 4, 0, 2, 6, 1,
+        //5, 8, 1, 0, 1, 9, 8, 4, 8, 0,
+    };
+    enum { NUM_DATA = sizeof(data) / sizeof(data[0]) };
+
+    int data_copy[NUM_DATA];
+    for (int end = 1; end < NUM_DATA; end++)
+    {
+        memcpy(data_copy, data, NUM_DATA * sizeof(data[0]));
+        dump_data("Unsorted", data_copy, 0, end);
+        quicksort(data_copy, 0, end);
+        dump_data("PostSort", data_copy, 0, end);
+        check_sorted(data_copy, 0, end);
+    }
+
+    return 0;
+}
